Refactor diagnostic_emitter.h into multiple files.
